Complete Project

+++ main.py
@@ -5,20 +5,31 @@ from flask_migrate import Migrate
 from werkzeug.security import generate_password_hash, check_password_hash
 from datetime import date
 from flask_login import UserMixin, login_user, logout_user, login_required, LoginManager, current_user
-from webforms import LoginForm, PostForm, PasswordForm, UserForm, NamerForm
+from webforms import LoginForm, PostForm, PasswordForm, UserForm, NamerForm, SearchForm
+import os
+from dotenv import load_dotenv
+from flask_ckeditor import CKEditor
+from werkzeug.utils import secure_filename
+import uuid as uuid
 
-
+load_dotenv(f"{os.getcwd()}/{'.env'}")
+WTF_SECRET_KEY = os.environ.get("WTF_SECRET_KEY")
 
 # Create Flask Instance
 app = Flask(__name__)
+ckeditor = CKEditor(app)
 # Add database (sqlite)
 # app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///users.db'
 # Add database (MySQL DB)
 # app.config['SQLALCHEMY_DATABASE_URI'] = 'mysql://username:password@localdb
 app.config['SQLALCHEMY_DATABASE_URI'] = 'mysql+pymysql://root:password123@localhost/users'
 # Secret key for WTF forms
-app.config["SECRET_KEY"] = "my super secret key"
+app.config["SECRET_KEY"] = WTF_SECRET_KEY
 # initialize The database
+
+UPLOAD_FOLDER = 'static/images/'
+app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
+
 db = SQLAlchemy(app)
 migrate = Migrate(app, db)
 
@@ -31,6 +42,42 @@ login_manager.login_view = 'login'
 @login_manager.user_loader
 def load_user(user_id):
     return Users.query.get(int(user_id))
+
+
+# Create Admin Page
+@app.route("/admin")
+@login_required
+def admin():
+    id = current_user.id
+    if id == 36:
+        return render_template("admin.html")
+    else:
+        flash("Sorry! You must have admin privilege to access this page!")
+        return redirect(url_for('dashboard'))
+
+
+# Pass form to navbar
+@app.context_processor
+def base():
+    form = SearchForm()
+    return dict(form=form)
+
+
+#  Create Search Function
+@app.route('/search', methods=["GET", "POST"])
+def search():
+    form = SearchForm()
+    posts = Posts.query
+    if form.validate_on_submit():
+        # Get data from submitted form
+        post_searched = form.searched.data
+        # Query the database
+        posts = posts.filter(Posts.content.like('%' + post_searched + '%'))
+        posts_found = posts.order_by(Posts.title).all()
+        return render_template("search.html",
+                               form=form,
+                               searched=post_searched,
+                               posts=posts_found)
 
 
 # create login page
@@ -62,30 +109,85 @@ def logout():
     flash("You Have Benn logged out. Thanks for stopping by ")
     return redirect(url_for('login'))
 
+
 # create dashboard page
 @app.route("/dashboard", methods=["GET", "POST"])
 @login_required
 def dashboard():
     form = UserForm()
-    return render_template('dashboard.html', form=form)
+    id = current_user.id
+    name_to_update = Users.query.get_or_404(id)
+    if request.method == "POST":
+        name_to_update.name = request.form['name']
+        name_to_update.email = request.form['email']
+        name_to_update.favorite_color = request.form['favorite_color']
+        name_to_update.username = request.form['username']
+        name_to_update.about_author = request.form['about_author']
+
+        #  Check for profile pic
+        if request.files['profile_pic']:
+            name_to_update.profile_pic = request.files['profile_pic']
+            # Grab image name
+            pic_filename = secure_filename(name_to_update.profile_pic.filename)
+            # Set UUID
+            pic_name = str(uuid.uuid1()) + "_" + pic_filename
+            # Save That Image
+            saver = request.files['profile_pic']
+            # Change it to a string to save to db
+            name_to_update.profile_pic = pic_name
+            try:
+                db.session.commit()
+                saver.save(os.path.join(UPLOAD_FOLDER, pic_name))
+                flash("User Updated Successfully")
+                return render_template("dashboard.html",
+                                       form=form,
+                                       name_to_update=name_to_update,
+                                       id=id)
+
+            except:
+                flash("Error! Looks like there was an error...")
+                return render_template("dashboard.html",
+                                       form=form,
+                                       name_to_update=name_to_update,
+                                       id=id)
+
+        else:
+            db.session.commit()
+            flash("User Updated Successfully")
+            return render_template("dashboard.html",
+                                   form=form,
+                                   name_to_update=name_to_update,
+                                   id=id)
+
+    else:
+        return render_template("dashboard.html",
+                               form=form,
+                               name_to_update=name_to_update,
+                               id=id)
 
 
 @app.route("/posts/delete/<int:id>")
+@login_required
 def delete_post(id):
     post_to_delete = Posts.query.get_or_404(id)
-    try:
-        db.session.delete(post_to_delete)
-        db.session.commit()
-        flash("Blog Post was deleted successfully")
+    id = current_user.id
+    if id == current_user.id or id == 36:
+        try:
+            db.session.delete(post_to_delete)
+            db.session.commit()
+            flash("Blog Post was deleted successfully")
+            all_posts = Posts.query.order_by(Posts.date_posted)
+            return render_template("posts.html", all_posts=all_posts)
+
+        except:
+            flash("Whoops, something went wrong. Try again")
+            all_posts = Posts.query.order_by(Posts.date_posted)
+            return render_template("posts.html",
+                                   all_posts=all_posts)
+    else:
+        flash("You Are Not Authorized To Delete That Post")
         all_posts = Posts.query.order_by(Posts.date_posted)
         return render_template("posts.html", all_posts=all_posts)
-
-    except:
-
-        flash("Whoops, something went wrong. Try again")
-        all_posts = Posts.query.order_by(Posts.date_posted)
-        return render_template("posts.html",
-                               all_posts=all_posts)
 
 
 @app.route("/posts")
@@ -103,36 +205,43 @@ def post(id):
 
 
 @app.route("/posts/edit/<int:id>", methods=["GET", "POST"])
+@login_required
 def edit_post(id):
     post = Posts.query.get_or_404(id)
     form = PostForm()
     if form.validate_on_submit():
+        poster = current_user.id
         post.title = form.title.data
-        post.author = form.author.data
         post.slug = form.slug.data
         post.content = form.content.data
+        poster_id = poster
         # Update database
         db.session.add(post)
         db.session.commit()
         flash("Post has been updated")
         return redirect(url_for('post', id=post.id))
-    form.title.data = post.title
-    form.author.data = post.author
-    form.slug.data = post.slug
-    form.content.data = post.content
-    return render_template("edit_post.html", form=form)
+    if current_user.id == post.poster_id or current_user.id == 36:
+        form.title.data = post.title
+        form.slug.data = post.slug
+        form.content.data = post.content
+        return render_template("edit_post.html", form=form)
+    else:
+        flash("You Aren't Authorizes To Edit This Post")
+        all_posts = Posts.query.order_by(Posts.date_posted)
+        return render_template("posts.html", all_posts=all_posts)
 
 
 # Add Post Page
-@app.route("/add-post", methods=["GET", "POST"])
+@app.route("/add_post", methods=["GET", "POST"])
 def add_post():
     form = PostForm()
     if form.validate_on_submit():
+        poster = current_user.id
         post = Posts(
             title=form.title.data,
             content=form.content.data,
-            author=form.author.data,
-            slug=form.slug.data
+            slug=form.slug.data,
+            poster_id=poster
         )
         # Add to database
         db.session.add(post)
@@ -141,7 +250,6 @@ def add_post():
         # Clear the form
         form.title.data = ''
         form.content.data = ''
-        form.author.data = ''
         form.slug.data = ''
         flash("Post added successfully")
     our_posts = Posts.query.order_by(Posts.date_posted)
@@ -164,6 +272,7 @@ def get_current_date():
 
 # Update Database Record
 @app.route("/update/<int:id>", methods=["GET", "POST"])
+@login_required
 def update(id):
     form = UserForm()
     name_to_update = Users.query.get_or_404(id)
@@ -172,6 +281,8 @@ def update(id):
         name_to_update.email = request.form['email']
         name_to_update.favorite_color = request.form['favorite_color']
         name_to_update.username = request.form['username']
+        name_to_update.about_author = request.form['about_author']
+
         try:
             db.session.commit()
             flash("User Updated Successfully")
@@ -179,6 +290,7 @@ def update(id):
             name_to_update.email = ''
             name_to_update.favorite_color = ''
             name_to_update.username = ''
+            name_to_update.about_author = ''
             return render_template("update.html",
                                    form=form,
                                    name_to_update=name_to_update,
@@ -194,6 +306,7 @@ def update(id):
         form.name.data = ''
         form.email.data = ''
         form.favorite_color.data = ''
+        form.about_author.data = ''
         return render_template("update.html",
                                form=form,
                                name_to_update=name_to_update,
@@ -226,6 +339,7 @@ def add_user():
                          username=form.username.data,
                          email=form.email.data,
                          favorite_color=form.favorite_color.data,
+                         about_author=form.about_author.data,
                          password_hash=hashed_pw)
             db.session.add(user)
             db.session.commit()
@@ -234,6 +348,7 @@ def add_user():
         form.username.data = ''
         form.email.data = ''
         form.favorite_color.data = ''
+        form.about_author.data = ''
         form.password_hash.data = ''
         flash("User added successfully")
     our_users = Users.query.order_by(Users.date_added)
@@ -245,27 +360,32 @@ def add_user():
 
 
 @app.route("/delete/<int:id>")
+@login_required
 def delete(id):
-    form = UserForm()
-    name = None
-    user_to_delete = Users.query.get_or_404(id)
-    try:
-        db.session.delete(user_to_delete)
-        db.session.commit()
-        flash("User deleted successfully")
+    if id == current_user.id:
+        form = UserForm()
+        name = None
+        user_to_delete = Users.query.get_or_404(id)
+        try:
+            db.session.delete(user_to_delete)
+            db.session.commit()
+            flash("User deleted successfully")
 
-        our_users = Users.query.order_by(Users.date_added)
-        return render_template("add_user.html",
-                               form=form,
-                               name=name,
-                               our_users=our_users)
-    except:
-        flash("Whoops, there was a problem deleting user, try again.")
-        our_users = Users.query.order_by(Users.date_added)
-        return render_template("add_user.html",
-                               form=form,
-                               name=name,
-                               our_users=our_users)
+            our_users = Users.query.order_by(Users.date_added)
+            return render_template("add_user.html",
+                                   form=form,
+                                   name=name,
+                                   our_users=our_users)
+        except:
+            flash("Whoops, there was a problem deleting user, try again.")
+            our_users = Users.query.order_by(Users.date_added)
+            return render_template("add_user.html",
+                                   form=form,
+                                   name=name,
+                                   our_users=our_users)
+    else:
+        flash("Sorry, you can't delete that user")
+        return redirect(url_for('dashboard'))
 
 
 @app.route("/")
@@ -354,22 +474,27 @@ class Posts(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     title = db.Column(db.String(255))
     content = db.Column(db.Text)
-    author = db.Column(db.String(120))
+    # author = db.Column(db.String(120))
     date_posted = db.Column(db.DateTime, default=datetime.utcnow)
     slug = db.Column(db.String(255))
+    #     Create Foreign Key to Link Users(refer to primary key of the user)
+    poster_id = db.Column(db.Integer, db.ForeignKey('users.id'))
+
 
 # Create DB model
-
-
 class Users(db.Model, UserMixin):
     id = db.Column(db.Integer, primary_key=True)
     username = db.Column(db.String(20), nullable=False, unique=True)
     name = db.Column(db.String(200), nullable=False)
     email = db.Column(db.String(120), nullable=False, unique=True)
     favorite_color = db.Column(db.String(120))
+    about_author = db.Column(db.Text(500), nullable=True)
     date_added = db.Column(db.DateTime, default=datetime.utcnow)
+    profile_pic = db.Column(db.String(120), nullable=True)
     #  Create password
     password_hash = db.Column(db.String(120))
+    #  User can have many posts
+    posts = db.relationship('Posts', backref='poster')
 
     @property
     def password(self):
@@ -387,8 +512,7 @@ class Users(db.Model, UserMixin):
         return '<Name %r>' % self.name
 
 
-
-
-
 if __name__ == "__main__":
     app.run(debug=True)
+
+
